Don't depend on libc inside a test
When the rustc-dev component is installed there will be two versions of
libc in the sysroot. This would result in an error as rustc doesn't know
which version to pick. In this case we don't actually depend on the libc
crate, so we can just remove the extern crate libc.